+++ libs/component/confirm-input/src/lib/component-confirm-input.module.ts
@@ -7,4 +7,4 @@ import { ConfirmInputDirective } from './confirm-input.directive';
   declarations: [ConfirmInputDirective],
   exports: [ConfirmInputDirective]
 })
-export class ComponentConfirmInputModule { }
+export class ComponentConfirmInputModule {}