+++ tools/schematics/ng-packagr-config-dest/index.ts
@@ -1,14 +1,22 @@
-import { chain, Rule, SchematicContext, Tree } from '@angular-devkit/schematics';
+import {
+  chain,
+  Rule,
+  SchematicContext,
+  Tree
+} from '@angular-devkit/schematics';
 import { PathFragment } from '@angular-devkit/core';
 
 const destination = './dist/';
 
-export default function (schema: any): Rule {
+export default function(schema: any): Rule {
   return chain([
-    function (tree: Tree, _context: SchematicContext) {
+    function(tree: Tree, _context: SchematicContext) {
       let path: string;
 
-      if (schema.path.startsWith('libs/') || schema.path.startsWith('./libs/')) {
+      if (
+        schema.path.startsWith('libs/') ||
+        schema.path.startsWith('./libs/')
+      ) {
         path = schema.path;
       } else {
         path = `libs/${schema.path}`;